+++ game/guitargraph.hh
@@ -86,7 +86,7 @@ class GuitarGraph: public InstrumentGraph {
 	Texture m_flame_godmode;
 	Surface m_tap; /// image for 2d HOPO note cap
 	Surface m_neckglow; /// image for the glow from the bottom of the neck
-	glmath::vec4 m_neckglowColor;
+	glmath::dvec4 m_neckglowColor;
 	Object3d m_fretObj; /// 3d object for regular note
 	Object3d m_tappableObj; /// 3d object for the HOPO note cap
 	std::vector<std::string> m_samples; /// sound effects
